Pull out sprint summary and aggregate reporting into reports submodule
in preparation for generating reports that can be stored and queried from the database.

This is mainly a performance optimisation. It will mean that reports can
be generated by asynchronous jobs, rather than on the fly in the frontend.
The frontend can then focus fully on plotting.

Since the reports will be stored in mongo, the header names were changed
to an undescore separated long form. There is now a separation between
data format and display format, so the sprint view needed updating.